refactor and add markdown to test
