+++ interaction/commands/register.py
